Add ngx-dropzone to Third Pary Components (#351)
Ngx-dropzone is a highly configurable Angular component to catch file drops or selections with a sleek default styling. 
You can also set options like the allowed file types or maximum sizes and provide a custom styled `div` which will use your styles but keeps the behavior and property bindings of the component.